+++ src/augmentation_class.cpp
@@ -6,17 +6,19 @@ using namespace arma;
 
 Augmentation::Augmentation(
   Data& dat,
-  const Rcpp::List& compute_options
+  const Rcpp::List& compute_options,
+  const Rcpp::List& model_options
 ) :
   save_aug { compute_options["save_aug"] },
   aug_thinning { compute_options["aug_thinning"] },
-  swap_leap { compute_options["swap_leap"] },
-  aug_method ( compute_options["aug_method"] ),
-  pseudo_aug_metric ( compute_options["pseudo_aug_metric"] ),
-  pseudo_aug_distance {
-    aug_method == "uniform" ? nullptr : choose_distance_function(pseudo_aug_metric)
-  },
-  log_aug_prob { zeros(dat.n_assessors) } {
+  error_model(model_options["error_model"]),
+  partial_aug_prop {
+    choose_partial_proposal(compute_options["aug_method"],
+                            compute_options["pseudo_aug_metric"]) },
+  pairwise_aug_prop {
+    choose_pairwise_proposal(error_model, compute_options["swap_leap"])
+  }
+   {
     if(dat.any_missing){
       dat.rankings = initialize_missing_ranks(dat.rankings, dat.missing_indicator);
     }
@@ -28,45 +30,37 @@ Augmentation::Augmentation(
       augmented_data.slice(0) = dat.rankings;
     }}
 
-void Augmentation::augment_pairwise(
-    Data& dat,
-    const Parameters& pars,
-    const Clustering& clus,
-    const std::unique_ptr<Distance>& distfun,
-    const std::unique_ptr<ProposalDistribution>& prop
-){
-  if(!dat.augpair) return;
-  for(size_t i = 0; i < dat.n_assessors; ++i) {
-    RankProposal rp = prop->propose(
-        dat.rankings.col(i), dat.items_above[i], dat.items_below[i]);
-
-    double u = std::log(R::runif(0, 1));
-    int cluster = clus.current_cluster_assignment(i);
-    double newdist = distfun->d(rp.rankings, pars.rho_old.col(cluster), rp.mutated_items);
-    double olddist = distfun->d(dat.rankings.col(i), pars.rho_old.col(cluster), rp.mutated_items);
-    double ratio = -pars.alpha_old(cluster) / dat.n_items * (newdist - olddist);
-
-    if(pars.error_model != "none") {
-      ratio += rp.g_diff * std::log(pars.theta_current / (1 - pars.theta_current));
-    }
-    if(ratio > u) dat.rankings.col(i) = rp.rankings;
-  }
-}
-
 void Augmentation::update_missing_ranks(
     Data& dat,
     const Clustering& clus,
     const Parameters& pars,
     const std::unique_ptr<Distance>& distfun) {
-  if(!dat.any_missing) return;
-
+  if(!dat.any_missing && !dat.augpair) return;
   for(size_t i = 0; i < dat.n_assessors; ++i){
     int cluster = clus.current_cluster_assignment(i);
-    dat.rankings.col(i) = make_new_augmentation(
-      dat.rankings.col(i), dat.missing_indicator.col(i),
-      pars.alpha_old(cluster), pars.rho_old.col(cluster),
-      distfun, pseudo_aug_distance,
-      log_aug_prob(i)
-    );
+    std::pair<vec, bool> aug{};
+    if(dat.any_missing) {
+      aug = make_new_augmentation(
+        dat.rankings.col(i), dat.missing_indicator.col(i),
+        pars.alpha_old(cluster), pars.rho_old.col(cluster),
+        distfun, partial_aug_prop);
+    } else if(dat.augpair) {
+      aug = make_new_augmentation(
+        dat.rankings.col(i), pars.alpha_old(cluster), pars.rho_old.col(cluster),
+        pars.theta_current, distfun, pairwise_aug_prop,
+        dat.items_above[i], dat.items_below[i], error_model);
+    }
+    if(pars.t > pars.burnin) aug_count++;
+    if(aug.second) {
+      dat.rankings.col(i) = aug.first;
+      if(pars.t > pars.burnin) aug_acceptance++;
+    }
+  }
+}
+
+void Augmentation::save_augmented_data(const Data& dat, const Parameters& pars) {
+  if(save_aug & (pars.t % aug_thinning == 0)){
+    ++aug_index;
+    augmented_data.slice(aug_index) = dat.rankings;
   }
 }