+++ src/classes.h
@@ -1,11 +1,9 @@
 #pragma once
 #include <RcppArmadillo.h>
+#include "typedefs.h"
 #include "partition_functions.h"
 #include "distances.h"
 #include "rank_proposal.h"
-
-using doubly_nested = std::vector<std::vector<unsigned int>>;
-using triply_nested = std::vector<doubly_nested>;
 
 struct Data {
   Data(const Rcpp::List& data);
@@ -15,8 +13,8 @@ struct Data {
   unsigned int n_assessors;
   const unsigned int n_items;
   arma::vec observation_frequency;
-  const triply_nested items_above{};
-  const triply_nested items_below{};
+  triply_nested items_above{};
+  triply_nested items_below{};
   const bool any_missing;
   const bool augpair;
   arma::umat missing_indicator;
@@ -46,8 +44,7 @@ struct Parameters {
   void update_shape(const Data& dat, const Priors& priors);
   void update_rho(const Data& dat,
                   const arma::uvec& cluster_assignment,
-                  const std::unique_ptr<Distance>& distfun,
-                  const std::unique_ptr<ProposalDistribution>& prop);
+                  const std::unique_ptr<Distance>& distfun);
 
   void update_alpha(
       const Data& dat,
@@ -68,19 +65,18 @@ struct Parameters {
   double theta_current;
   const unsigned int n_clusters;
   const unsigned int nmc;
-  const std::string error_model;
   const int alpha_jump;
-  const int leap_size;
-  const std::string rho_proposal_option;
   const std::string metric;
-  int burnin{};
+  size_t burnin{};
   size_t t{};
   size_t alpha_index{};
   size_t rho_index{};
 
 private:
+  const std::unique_ptr<RhoProposal> rho_proposal_function;
   const double alpha_prop_sd;
   const int rho_thinning;
+  const std::string error_model;
 };
 
 struct Clustering {
@@ -95,10 +91,12 @@ struct Clustering {
   void update_wcd(const int t);
   void update_dist_mat(const Data& dat, const Parameters& pars,
                        const std::unique_ptr<Distance>& distfun);
+  void save_cluster_parameters(size_t t);
 
   arma::vec current_cluster_probs;
   const bool clustering;
   const unsigned int clus_thinning;
+  size_t cluster_assignment_index{};
 
 private:
   const arma::uvec index;
@@ -115,15 +113,9 @@ public:
 };
 
 struct Augmentation {
-  Augmentation(Data& dat, const Rcpp::List& compute_options);
+  Augmentation(Data& dat, const Rcpp::List& compute_options,
+               const Rcpp::List& model_options);
   ~Augmentation() = default;
-
-  void augment_pairwise(
-      Data& dat,
-      const Parameters& pars,
-      const Clustering& clus,
-      const std::unique_ptr<Distance>& distfun,
-      const std::unique_ptr<ProposalDistribution>& prop);
 
   void update_missing_ranks(
       Data& dat,
@@ -131,16 +123,17 @@ struct Augmentation {
       const Parameters& pars,
       const std::unique_ptr<Distance>& distfun
   );
-
+  void save_augmented_data(const Data& dat, const Parameters& pars);
   const bool save_aug;
   const unsigned int aug_thinning;
   arma::cube augmented_data;
-  const unsigned int swap_leap;
+  size_t aug_index{};
+  double aug_acceptance{};
+  int aug_count{};
 
 private:
-  const std::string aug_method;
-  const std::string pseudo_aug_metric;
-  const std::unique_ptr<Distance> pseudo_aug_distance;
-  arma::vec log_aug_prob;
+  const std::string error_model;
+  const std::unique_ptr<PartialProposal> partial_aug_prop;
+  const std::unique_ptr<PairwiseProposal> pairwise_aug_prop;
 };