+++ src/run_smc.cpp
@@ -32,23 +32,20 @@ Rcpp::List  run_smc(
   auto pfun = choose_partition_function(
     dat.n_items, pars.metric, pfun_values, pfun_estimate);
   auto distfun = choose_distance_function(pars.metric);
-  auto rho_proposal = choose_rank_proposal(
-    pars.rho_proposal_option, pars.leap_size);
 
-  auto T{new_data.size()};
+  size_t T = new_data.size();
   for(size_t t{}; t < T; t++) {
     dat.update(new_data[t]);
     particle_vectors[t + 1] = augment_particles(particle_vectors[t], dat);
     aug.reweight(particle_vectors[t + 1], dat, pfun, distfun);
     pars.resample(particle_vectors[t + 1]);
-
     par_for_each(
       particle_vectors[t + 1].begin(), particle_vectors[t + 1].end(),
       [&pars, &dat, &pris, &aug, distfun = std::ref(distfun),
-       pfun = std::ref(pfun), rho_proposal = std::ref(rho_proposal)]
+       pfun = std::ref(pfun)]
       (Particle& p) {
          for(size_t i{}; i < pars.mcmc_steps; i++) {
-           pars.update_rho(p, dat, distfun, rho_proposal);
+           pars.update_rho(p, dat, distfun);
            pars.update_alpha(p, dat, pfun, distfun, pris);
            aug.update_missing_ranks(p, dat, distfun);
          }
@@ -64,6 +61,5 @@ Rcpp::List  run_smc(
     Rcpp::Named("data") = dat.wrapup(),
     Rcpp::Named("acceptance_ratios") = compute_particle_acceptance(particle_vectors, pars.mcmc_steps)
   );
-
   return particle_history;
 }