+++ src/setdiff.h
@@ -1,0 +1,3 @@
+#pragma once
+#include <RcppArmadillo.h>
+arma::vec setdiff(const arma::vec& x, const arma::vec& y);